+++ src/api/users.js
@@ -1,11 +1,15 @@
-import { Sdk } from '@/sdk'
+import { ApiCallerFactory } from '@/api-caller-factory'
 
 export default {
 
   async getAccountIdByEmail (email) {
     try {
-      const { data } = await Sdk.horizon.public.getAccountIdByEmail(email)
-      return data.accountId
+      const { data } = await ApiCallerFactory
+        .createCallerInstance()
+        .get('/identities', {
+          filter: { email }
+        })
+      return ((data || [])[0] || {}).address
     } catch (error) {
       if (error.httpStatus === 404) {
         return ''